Total refactoring to make the project OSGi-based.
