+++ skinny-web-api/bnd.bnd
@@ -1,0 +1,9 @@
+Bundle-Name: ${project.name}
+Bundle-SymbolicName: com.liferay.skinny.api
+Bundle-Version: ${project.version}
+Export-Package:\
+	com.liferay.skinny.exception,\
+	com.liferay.skinny.model,\
+	com.liferay.skinny.service,\
+	com.liferay.skinny.service.persistence
+-includeresource: META-INF/service.xml=../skinny-web-service/service.xml
