+++ skinny-web-api/src/main/java/com/liferay/skinny/model/SkinnyDDLRecord.java
