+++ skinny-web-api/src/main/java/com/liferay/skinny/model/SkinnyJournalArticle.java
