+++ skinny-web-api/src/main/java/com/liferay/skinny/model/SkinnyJournalArticleVersionMetadata.java
