+++ skinny-web-api/src/main/java/com/liferay/skinny/service/SkinnyService.java
@@ -1,0 +1,94 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.skinny.service;
+
+import aQute.bnd.annotation.ProviderType;
+
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.exception.SystemException;
+import com.liferay.portal.kernel.jsonwebservice.JSONWebService;
+import com.liferay.portal.kernel.security.access.control.AccessControlled;
+import com.liferay.portal.kernel.service.BaseService;
+import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
+import com.liferay.portal.kernel.transaction.Isolation;
+import com.liferay.portal.kernel.transaction.Propagation;
+import com.liferay.portal.kernel.transaction.Transactional;
+
+import com.liferay.skinny.model.SkinnyDDLRecord;
+import com.liferay.skinny.model.SkinnyJournalArticle;
+import com.liferay.skinny.model.SkinnyJournalArticleVersionMetadata;
+
+import java.util.List;
+
+/**
+ * Provides the remote service interface for Skinny. Methods of this
+ * service are expected to have security checks based on the propagated JAAS
+ * credentials because this service can be accessed remotely.
+ *
+ * @author Brian Wing Shun Chan
+ * @see SkinnyServiceUtil
+ * @see com.liferay.skinny.service.base.SkinnyServiceBaseImpl
+ * @see com.liferay.skinny.service.impl.SkinnyServiceImpl
+ * @generated
+ */
+@AccessControlled
+@JSONWebService
+@OSGiBeanProperties(property =  {
+	"json.web.service.context.name=skinny", "json.web.service.context.path=Skinny"}, service = SkinnyService.class)
+@ProviderType
+@Transactional(isolation = Isolation.PORTAL, rollbackFor =  {
+	PortalException.class, SystemException.class})
+public interface SkinnyService extends BaseService {
+	/*
+	 * NOTE FOR DEVELOPERS:
+	 *
+	 * Never modify or reference this interface directly. Always use {@link SkinnyServiceUtil} to access the skinny remote service. Add custom service methods to {@link com.liferay.skinny.service.impl.SkinnyServiceImpl} and rerun ServiceBuilder to automatically copy the method declarations to this interface.
+	 */
+
+	/**
+	* Returns the OSGi service identifier.
+	*
+	* @return the OSGi service identifier
+	*/
+	public java.lang.String getOSGiServiceIdentifier();
+
+	@AccessControlled(guestAccessEnabled = true)
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public List<SkinnyDDLRecord> getSkinnyDDLRecords(long ddlRecordSetId)
+		throws java.lang.Exception;
+
+	@AccessControlled(guestAccessEnabled = true)
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public SkinnyJournalArticle getSkinnyJournalArticle(long groupId,
+		java.lang.String articleId, int status, java.lang.String locale)
+		throws java.lang.Exception;
+
+	@AccessControlled(guestAccessEnabled = true)
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public SkinnyJournalArticle getSkinnyJournalArticleByVersion(long groupId,
+		java.lang.String articleId, java.lang.String version,
+		java.lang.String locale) throws java.lang.Exception;
+
+	@AccessControlled(guestAccessEnabled = true)
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public List<SkinnyJournalArticle> getSkinnyJournalArticles(long companyId,
+		java.lang.String groupName, long ddmStructureId, java.lang.String locale)
+		throws java.lang.Exception;
+
+	@AccessControlled(guestAccessEnabled = true)
+	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
+	public List<SkinnyJournalArticleVersionMetadata> getSkinnyJournalArticleVersions(
+		long groupId, java.lang.String articleId) throws java.lang.Exception;
+}
