+++ skinny-web-service/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: ${project.name}
+Bundle-SymbolicName: com.liferay.skinny.service
+Bundle-Version: ${project.version}
+Liferay-Require-SchemaVersion: 1.0.0
+Liferay-Service: true
+-plugin.service: com.liferay.ant.bnd.service.ServiceAnalyzerPlugin
+-plugin.spring: com.liferay.ant.bnd.spring.SpringDependencyAnalyzerPlugin
