+++ skinny-web-service/pom.xml
@@ -1,0 +1,100 @@
+<?xml version="1.0"?>
+
+<project
+	xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+>
+	<modelVersion>4.0.0</modelVersion>
+	<artifactId>skinny-web-service</artifactId>
+	<name>Skinny JSON WebService :: Service</name>
+	<version>1.1-SNAPSHOT</version>
+	<parent>
+		<groupId>se.vgregion.reklistan.liferay</groupId>
+		<artifactId>skinny-web</artifactId>
+		<version>1.1-SNAPSHOT</version>
+	</parent>
+	<dependencies>
+		<dependency>
+			<groupId>com.liferay</groupId>
+			<artifactId>com.liferay.portal.spring.extender</artifactId>
+			<version>2.1.3</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>com.liferay</groupId>
+			<artifactId>com.liferay.osgi.service.tracker.collections</artifactId>
+			<version>2.0.3</version>
+			<scope>provided</scope>
+		</dependency>
+        <dependency>
+            <groupId>se.vgregion.reklistan.liferay</groupId>
+            <artifactId>skinny-web-api</artifactId>
+            <version>1.1-SNAPSHOT</version>
+        </dependency>
+		<dependency>
+			<groupId>com.liferay.portal</groupId>
+			<artifactId>com.liferay.portal.kernel</artifactId>
+			<version>2.49.0</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
+            <groupId>com.liferay</groupId>
+            <artifactId>com.liferay.dynamic.data.lists.api</artifactId>
+            <version>2.1.7</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.liferay</groupId>
+            <artifactId>com.liferay.dynamic.data.mapping.api</artifactId>
+            <version>3.7.2</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.liferay</groupId>
+            <artifactId>com.liferay.journal.api</artifactId>
+            <version>2.12.1</version>
+            <scope>provided</scope>
+        </dependency>
+		<dependency>
+			<groupId>biz.aQute.bnd</groupId>
+			<artifactId>biz.aQute.bnd.annotation</artifactId>
+			<version>3.5.0</version>
+			<scope>provided</scope>
+		</dependency>
+	</dependencies>
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>com.liferay</groupId>
+				<artifactId>com.liferay.portal.tools.service.builder</artifactId>
+				<version>1.0.178</version>
+				<configuration>
+					<apiDirName>../skinny-web-api/src/main/java</apiDirName>
+					<autoNamespaceTables>true</autoNamespaceTables>
+					<buildNumberIncrement>true</buildNumberIncrement>
+					<hbmFileName>src/main/resources/META-INF/module-hbm.xml</hbmFileName>
+					<implDirName>src/main/java</implDirName>
+					<mergeModelHintsConfigs>src/main/resources/META-INF/portlet-model-hints.xml</mergeModelHintsConfigs>
+					<modelHintsFileName>src/main/resources/META-INF/portlet-model-hints.xml</modelHintsFileName>
+					<osgiModule>true</osgiModule>
+					<propsUtil>com.liferay.skinny.service.util.ServiceProps</propsUtil>
+					<resourcesDirName>src/main/resources</resourcesDirName>
+					<springFileName>src/main/resources/META-INF/spring/module-spring.xml</springFileName>
+					<sqlDirName>src/main/resources/META-INF/sql</sqlDirName>
+					<sqlFileName>tables.sql</sqlFileName>
+				</configuration>
+			</plugin>
+			<plugin>
+                <groupId>com.liferay</groupId>
+                <artifactId>com.liferay.portal.tools.wsdd.builder</artifactId>
+                <version>1.0.10</version>
+                <configuration>
+                    <inputFileName>service.xml</inputFileName>
+                    <outputDirName>src/main/java</outputDirName>
+                    <serverConfigFileName>src/main/resources/server-config.wsdd</serverConfigFileName>
+                </configuration>
+            </plugin>
+		</plugins>
+	</build>
+</project>
