+++ skinny-web-service/src/main/resources/META-INF/portlet-model-hints.xml
