+++ skinny-web-service/src/main/resources/META-INF/spring/module-spring.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+
+<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" default-destroy-method="destroy" default-init-method="afterPropertiesSet" xsi:schemaLocation=" http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+	<bean class="com.liferay.skinny.service.impl.SkinnyServiceImpl" id="com.liferay.skinny.service.SkinnyService" />
+</beans>
