+++ skinny-web-service/src/main/resources/service.properties
@@ -1,0 +1,17 @@
+##
+## Properties Override
+##
+
+    #
+    # Specify where to get the overridden properties. Updates should not be made
+    # on this file but on the overridden version of this file.
+    #
+    include-and-override=${base.path}/service-ext.properties
+
+##
+## Build
+##
+
+    build.namespace=Skinny
+    build.number=9
+    build.date=1513935916682
