--- src/main/java/com/liferay/skinny/service/ClpSerializer.java
@@ -1,189 +1,0 @@
-package com.liferay.skinny.service;
-
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayInputStream;
-import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.ClassLoaderObjectInputStream;
-import com.liferay.portal.kernel.util.PropsUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.model.BaseModel;
-
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-
-import java.lang.reflect.Method;
-
-import java.util.ArrayList;
-import java.util.List;
-
-
-public class ClpSerializer {
-    private static Log _log = LogFactoryUtil.getLog(ClpSerializer.class);
-    private static String _servletContextName;
-    private static boolean _useReflectionToTranslateThrowable = true;
-
-    public static String getServletContextName() {
-        if (Validator.isNotNull(_servletContextName)) {
-            return _servletContextName;
-        }
-
-        synchronized (ClpSerializer.class) {
-            if (Validator.isNotNull(_servletContextName)) {
-                return _servletContextName;
-            }
-
-            try {
-                ClassLoader classLoader = ClpSerializer.class.getClassLoader();
-
-                Class<?> portletPropsClass = classLoader.loadClass(
-                        "com.liferay.util.portlet.PortletProps");
-
-                Method getMethod = portletPropsClass.getMethod("get",
-                        new Class<?>[] { String.class });
-
-                String portletPropsServletContextName = (String) getMethod.invoke(null,
-                        "skinny-web-deployment-context");
-
-                if (Validator.isNotNull(portletPropsServletContextName)) {
-                    _servletContextName = portletPropsServletContextName;
-                }
-            } catch (Throwable t) {
-                if (_log.isInfoEnabled()) {
-                    _log.info(
-                        "Unable to locate deployment context from portlet properties");
-                }
-            }
-
-            if (Validator.isNull(_servletContextName)) {
-                try {
-                    String propsUtilServletContextName = PropsUtil.get(
-                            "skinny-web-deployment-context");
-
-                    if (Validator.isNotNull(propsUtilServletContextName)) {
-                        _servletContextName = propsUtilServletContextName;
-                    }
-                } catch (Throwable t) {
-                    if (_log.isInfoEnabled()) {
-                        _log.info(
-                            "Unable to locate deployment context from portal properties");
-                    }
-                }
-            }
-
-            if (Validator.isNull(_servletContextName)) {
-                _servletContextName = "skinny-web";
-            }
-
-            return _servletContextName;
-        }
-    }
-
-    public static Object translateInput(BaseModel<?> oldModel) {
-        return oldModel;
-    }
-
-    public static Object translateInput(List<Object> oldList) {
-        List<Object> newList = new ArrayList<Object>(oldList.size());
-
-        for (int i = 0; i < oldList.size(); i++) {
-            Object curObj = oldList.get(i);
-
-            newList.add(translateInput(curObj));
-        }
-
-        return newList;
-    }
-
-    public static Object translateInput(Object obj) {
-        if (obj instanceof BaseModel<?>) {
-            return translateInput((BaseModel<?>) obj);
-        } else if (obj instanceof List<?>) {
-            return translateInput((List<Object>) obj);
-        } else {
-            return obj;
-        }
-    }
-
-    public static Object translateOutput(BaseModel<?> oldModel) {
-        return oldModel;
-    }
-
-    public static Object translateOutput(List<Object> oldList) {
-        List<Object> newList = new ArrayList<Object>(oldList.size());
-
-        for (int i = 0; i < oldList.size(); i++) {
-            Object curObj = oldList.get(i);
-
-            newList.add(translateOutput(curObj));
-        }
-
-        return newList;
-    }
-
-    public static Object translateOutput(Object obj) {
-        if (obj instanceof BaseModel<?>) {
-            return translateOutput((BaseModel<?>) obj);
-        } else if (obj instanceof List<?>) {
-            return translateOutput((List<Object>) obj);
-        } else {
-            return obj;
-        }
-    }
-
-    public static Throwable translateThrowable(Throwable throwable) {
-        if (_useReflectionToTranslateThrowable) {
-            try {
-                UnsyncByteArrayOutputStream unsyncByteArrayOutputStream = new UnsyncByteArrayOutputStream();
-                ObjectOutputStream objectOutputStream = new ObjectOutputStream(unsyncByteArrayOutputStream);
-
-                objectOutputStream.writeObject(throwable);
-
-                objectOutputStream.flush();
-                objectOutputStream.close();
-
-                UnsyncByteArrayInputStream unsyncByteArrayInputStream = new UnsyncByteArrayInputStream(unsyncByteArrayOutputStream.unsafeGetByteArray(),
-                        0, unsyncByteArrayOutputStream.size());
-
-                Thread currentThread = Thread.currentThread();
-
-                ClassLoader contextClassLoader = currentThread.getContextClassLoader();
-
-                ObjectInputStream objectInputStream = new ClassLoaderObjectInputStream(unsyncByteArrayInputStream,
-                        contextClassLoader);
-
-                throwable = (Throwable) objectInputStream.readObject();
-
-                objectInputStream.close();
-
-                return throwable;
-            } catch (SecurityException se) {
-                if (_log.isInfoEnabled()) {
-                    _log.info("Do not use reflection to translate throwable");
-                }
-
-                _useReflectionToTranslateThrowable = false;
-            } catch (Throwable throwable2) {
-                _log.error(throwable2, throwable2);
-
-                return throwable2;
-            }
-        }
-
-        Class<?> clazz = throwable.getClass();
-
-        String className = clazz.getName();
-
-        if (className.equals(PortalException.class.getName())) {
-            return new PortalException();
-        }
-
-        if (className.equals(SystemException.class.getName())) {
-            return new SystemException();
-        }
-
-        return throwable;
-    }
-}
