--- src/main/java/com/liferay/skinny/service/SkinnyService.java
@@ -1,82 +1,0 @@
-package com.liferay.skinny.service;
-
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.jsonwebservice.JSONWebService;
-import com.liferay.portal.kernel.transaction.Isolation;
-import com.liferay.portal.kernel.transaction.Propagation;
-import com.liferay.portal.kernel.transaction.Transactional;
-import com.liferay.portal.security.ac.AccessControlled;
-import com.liferay.portal.service.BaseService;
-import com.liferay.portal.service.InvokableService;
-
-/**
- * Provides the remote service interface for Skinny. Methods of this
- * service are expected to have security checks based on the propagated JAAS
- * credentials because this service can be accessed remotely.
- *
- * @author Brian Wing Shun Chan
- * @see SkinnyServiceUtil
- * @see com.liferay.skinny.service.base.SkinnyServiceBaseImpl
- * @see com.liferay.skinny.service.impl.SkinnyServiceImpl
- * @generated
- */
-@AccessControlled
-@JSONWebService
-@Transactional(isolation = Isolation.PORTAL, rollbackFor =  {
-    PortalException.class, SystemException.class}
-)
-public interface SkinnyService extends BaseService, InvokableService {
-    /*
-     * NOTE FOR DEVELOPERS:
-     *
-     * Never modify or reference this interface directly. Always use {@link SkinnyServiceUtil} to access the skinny remote service. Add custom service methods to {@link com.liferay.skinny.service.impl.SkinnyServiceImpl} and rerun ServiceBuilder to automatically copy the method declarations to this interface.
-     */
-
-    /**
-    * Returns the Spring bean ID for this bean.
-    *
-    * @return the Spring bean ID for this bean
-    */
-    public java.lang.String getBeanIdentifier();
-
-    /**
-    * Sets the Spring bean ID for this bean.
-    *
-    * @param beanIdentifier the Spring bean ID for this bean
-    */
-    public void setBeanIdentifier(java.lang.String beanIdentifier);
-
-    @Override
-    public java.lang.Object invokeMethod(java.lang.String name,
-        java.lang.String[] parameterTypes, java.lang.Object[] arguments)
-        throws java.lang.Throwable;
-
-    @com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
-    @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-    public java.util.List<com.liferay.skinny.model.SkinnyDDLRecord> getSkinnyDDLRecords(
-        long ddlRecordSetId) throws java.lang.Exception;
-
-    @com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
-    @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-    public java.util.List<com.liferay.skinny.model.SkinnyJournalArticle> getSkinnyJournalArticles(
-        long companyId, java.lang.String groupName, long ddmStructureId,
-        java.lang.String locale) throws java.lang.Exception;
-
-    @com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
-    @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-    public com.liferay.skinny.model.SkinnyJournalArticle getSkinnyJournalArticle(
-        long groupId, java.lang.String articleId, int status,
-        java.lang.String locale) throws java.lang.Exception;
-
-    @com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
-    @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-    public java.util.List<com.liferay.skinny.model.SkinnyJournalArticleVersionMetadata> getSkinnyJournalArticleVersions(
-        long groupId, java.lang.String articleId) throws java.lang.Exception;
-
-    @com.liferay.portal.security.ac.AccessControlled(guestAccessEnabled = true)
-    @Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
-    public com.liferay.skinny.model.SkinnyJournalArticle getSkinnyJournalArticleByVersion(
-        long groupId, java.lang.String articleId, java.lang.String version,
-        java.lang.String locale) throws java.lang.Exception;
-}
