--- src/main/java/com/liferay/skinny/service/SkinnyServiceUtil.java
@@ -1,114 +1,0 @@
-package com.liferay.skinny.service;
-
-import com.liferay.portal.kernel.bean.PortletBeanLocatorUtil;
-import com.liferay.portal.kernel.util.ReferenceRegistry;
-import com.liferay.portal.service.InvokableService;
-
-/**
- * Provides the remote service utility for Skinny. This utility wraps
- * {@link com.liferay.skinny.service.impl.SkinnyServiceImpl} and is the
- * primary access point for service operations in application layer code running
- * on a remote server. Methods of this service are expected to have security
- * checks based on the propagated JAAS credentials because this service can be
- * accessed remotely.
- *
- * @author Brian Wing Shun Chan
- * @see SkinnyService
- * @see com.liferay.skinny.service.base.SkinnyServiceBaseImpl
- * @see com.liferay.skinny.service.impl.SkinnyServiceImpl
- * @generated
- */
-public class SkinnyServiceUtil {
-    private static SkinnyService _service;
-
-    /*
-     * NOTE FOR DEVELOPERS:
-     *
-     * Never modify this class directly. Add custom service methods to {@link com.liferay.skinny.service.impl.SkinnyServiceImpl} and rerun ServiceBuilder to regenerate this class.
-     */
-
-    /**
-    * Returns the Spring bean ID for this bean.
-    *
-    * @return the Spring bean ID for this bean
-    */
-    public static java.lang.String getBeanIdentifier() {
-        return getService().getBeanIdentifier();
-    }
-
-    /**
-    * Sets the Spring bean ID for this bean.
-    *
-    * @param beanIdentifier the Spring bean ID for this bean
-    */
-    public static void setBeanIdentifier(java.lang.String beanIdentifier) {
-        getService().setBeanIdentifier(beanIdentifier);
-    }
-
-    public static java.lang.Object invokeMethod(java.lang.String name,
-        java.lang.String[] parameterTypes, java.lang.Object[] arguments)
-        throws java.lang.Throwable {
-        return getService().invokeMethod(name, parameterTypes, arguments);
-    }
-
-    public static java.util.List<com.liferay.skinny.model.SkinnyDDLRecord> getSkinnyDDLRecords(
-        long ddlRecordSetId) throws java.lang.Exception {
-        return getService().getSkinnyDDLRecords(ddlRecordSetId);
-    }
-
-    public static java.util.List<com.liferay.skinny.model.SkinnyJournalArticle> getSkinnyJournalArticles(
-        long companyId, java.lang.String groupName, long ddmStructureId,
-        java.lang.String locale) throws java.lang.Exception {
-        return getService()
-                   .getSkinnyJournalArticles(companyId, groupName,
-            ddmStructureId, locale);
-    }
-
-    public static com.liferay.skinny.model.SkinnyJournalArticle getSkinnyJournalArticle(
-        long groupId, java.lang.String articleId, int status,
-        java.lang.String locale) throws java.lang.Exception {
-        return getService()
-                   .getSkinnyJournalArticle(groupId, articleId, status, locale);
-    }
-
-    public static java.util.List<com.liferay.skinny.model.SkinnyJournalArticleVersionMetadata> getSkinnyJournalArticleVersions(
-        long groupId, java.lang.String articleId) throws java.lang.Exception {
-        return getService().getSkinnyJournalArticleVersions(groupId, articleId);
-    }
-
-    public static com.liferay.skinny.model.SkinnyJournalArticle getSkinnyJournalArticleByVersion(
-        long groupId, java.lang.String articleId, java.lang.String version,
-        java.lang.String locale) throws java.lang.Exception {
-        return getService()
-                   .getSkinnyJournalArticleByVersion(groupId, articleId,
-            version, locale);
-    }
-
-    public static void clearService() {
-        _service = null;
-    }
-
-    public static SkinnyService getService() {
-        if (_service == null) {
-            InvokableService invokableService = (InvokableService) PortletBeanLocatorUtil.locate(ClpSerializer.getServletContextName(),
-                    SkinnyService.class.getName());
-
-            if (invokableService instanceof SkinnyService) {
-                _service = (SkinnyService) invokableService;
-            } else {
-                _service = new SkinnyServiceClp(invokableService);
-            }
-
-            ReferenceRegistry.registerReference(SkinnyServiceUtil.class,
-                "_service");
-        }
-
-        return _service;
-    }
-
-    /**
-     * @deprecated As of 6.2.0
-     */
-    public void setService(SkinnyService service) {
-    }
-}
