--- src/main/java/com/liferay/skinny/service/base/SkinnyServiceBaseImpl.java
@@ -1,487 +1,0 @@
-package com.liferay.skinny.service.base;
-
-import com.liferay.portal.kernel.bean.BeanReference;
-import com.liferay.portal.kernel.bean.IdentifiableBean;
-import com.liferay.portal.kernel.dao.jdbc.SqlUpdate;
-import com.liferay.portal.kernel.dao.jdbc.SqlUpdateFactoryUtil;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.util.InfrastructureUtil;
-import com.liferay.portal.service.BaseServiceImpl;
-import com.liferay.portal.service.persistence.GroupPersistence;
-import com.liferay.portal.service.persistence.UserPersistence;
-
-import com.liferay.portlet.dynamicdatalists.service.persistence.DDLRecordSetPersistence;
-import com.liferay.portlet.dynamicdatamapping.service.persistence.DDMStructurePersistence;
-import com.liferay.portlet.journal.service.persistence.JournalArticlePersistence;
-
-import com.liferay.skinny.service.SkinnyService;
-
-import javax.sql.DataSource;
-
-/**
- * Provides the base implementation for the skinny remote service.
- *
- * <p>
- * This implementation exists only as a container for the default service methods generated by ServiceBuilder. All custom service methods should be put in {@link com.liferay.skinny.service.impl.SkinnyServiceImpl}.
- * </p>
- *
- * @author Brian Wing Shun Chan
- * @see com.liferay.skinny.service.impl.SkinnyServiceImpl
- * @see com.liferay.skinny.service.SkinnyServiceUtil
- * @generated
- */
-public abstract class SkinnyServiceBaseImpl extends BaseServiceImpl
-    implements SkinnyService, IdentifiableBean {
-    @BeanReference(type = com.liferay.skinny.service.SkinnyService.class)
-    protected com.liferay.skinny.service.SkinnyService skinnyService;
-    @BeanReference(type = com.liferay.counter.service.CounterLocalService.class)
-    protected com.liferay.counter.service.CounterLocalService counterLocalService;
-    @BeanReference(type = com.liferay.portal.service.GroupLocalService.class)
-    protected com.liferay.portal.service.GroupLocalService groupLocalService;
-    @BeanReference(type = com.liferay.portal.service.GroupService.class)
-    protected com.liferay.portal.service.GroupService groupService;
-    @BeanReference(type = GroupPersistence.class)
-    protected GroupPersistence groupPersistence;
-    @BeanReference(type = com.liferay.portal.service.ResourceLocalService.class)
-    protected com.liferay.portal.service.ResourceLocalService resourceLocalService;
-    @BeanReference(type = com.liferay.portal.service.UserLocalService.class)
-    protected com.liferay.portal.service.UserLocalService userLocalService;
-    @BeanReference(type = com.liferay.portal.service.UserService.class)
-    protected com.liferay.portal.service.UserService userService;
-    @BeanReference(type = UserPersistence.class)
-    protected UserPersistence userPersistence;
-    @BeanReference(type = com.liferay.portlet.dynamicdatalists.service.DDLRecordSetLocalService.class)
-    protected com.liferay.portlet.dynamicdatalists.service.DDLRecordSetLocalService ddlRecordSetLocalService;
-    @BeanReference(type = com.liferay.portlet.dynamicdatalists.service.DDLRecordSetService.class)
-    protected com.liferay.portlet.dynamicdatalists.service.DDLRecordSetService ddlRecordSetService;
-    @BeanReference(type = DDLRecordSetPersistence.class)
-    protected DDLRecordSetPersistence ddlRecordSetPersistence;
-    @BeanReference(type = com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalService.class)
-    protected com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalService ddmStructureLocalService;
-    @BeanReference(type = com.liferay.portlet.dynamicdatamapping.service.DDMStructureService.class)
-    protected com.liferay.portlet.dynamicdatamapping.service.DDMStructureService ddmStructureService;
-    @BeanReference(type = DDMStructurePersistence.class)
-    protected DDMStructurePersistence ddmStructurePersistence;
-    @BeanReference(type = com.liferay.portlet.journal.service.JournalArticleLocalService.class)
-    protected com.liferay.portlet.journal.service.JournalArticleLocalService journalArticleLocalService;
-    @BeanReference(type = com.liferay.portlet.journal.service.JournalArticleService.class)
-    protected com.liferay.portlet.journal.service.JournalArticleService journalArticleService;
-    @BeanReference(type = JournalArticlePersistence.class)
-    protected JournalArticlePersistence journalArticlePersistence;
-    private String _beanIdentifier;
-    private ClassLoader _classLoader;
-    private SkinnyServiceClpInvoker _clpInvoker = new SkinnyServiceClpInvoker();
-
-    /*
-     * NOTE FOR DEVELOPERS:
-     *
-     * Never modify or reference this class directly. Always use {@link com.liferay.skinny.service.SkinnyServiceUtil} to access the skinny remote service.
-     */
-
-    /**
-     * Returns the skinny remote service.
-     *
-     * @return the skinny remote service
-     */
-    public com.liferay.skinny.service.SkinnyService getSkinnyService() {
-        return skinnyService;
-    }
-
-    /**
-     * Sets the skinny remote service.
-     *
-     * @param skinnyService the skinny remote service
-     */
-    public void setSkinnyService(
-        com.liferay.skinny.service.SkinnyService skinnyService) {
-        this.skinnyService = skinnyService;
-    }
-
-    /**
-     * Returns the counter local service.
-     *
-     * @return the counter local service
-     */
-    public com.liferay.counter.service.CounterLocalService getCounterLocalService() {
-        return counterLocalService;
-    }
-
-    /**
-     * Sets the counter local service.
-     *
-     * @param counterLocalService the counter local service
-     */
-    public void setCounterLocalService(
-        com.liferay.counter.service.CounterLocalService counterLocalService) {
-        this.counterLocalService = counterLocalService;
-    }
-
-    /**
-     * Returns the group local service.
-     *
-     * @return the group local service
-     */
-    public com.liferay.portal.service.GroupLocalService getGroupLocalService() {
-        return groupLocalService;
-    }
-
-    /**
-     * Sets the group local service.
-     *
-     * @param groupLocalService the group local service
-     */
-    public void setGroupLocalService(
-        com.liferay.portal.service.GroupLocalService groupLocalService) {
-        this.groupLocalService = groupLocalService;
-    }
-
-    /**
-     * Returns the group remote service.
-     *
-     * @return the group remote service
-     */
-    public com.liferay.portal.service.GroupService getGroupService() {
-        return groupService;
-    }
-
-    /**
-     * Sets the group remote service.
-     *
-     * @param groupService the group remote service
-     */
-    public void setGroupService(
-        com.liferay.portal.service.GroupService groupService) {
-        this.groupService = groupService;
-    }
-
-    /**
-     * Returns the group persistence.
-     *
-     * @return the group persistence
-     */
-    public GroupPersistence getGroupPersistence() {
-        return groupPersistence;
-    }
-
-    /**
-     * Sets the group persistence.
-     *
-     * @param groupPersistence the group persistence
-     */
-    public void setGroupPersistence(GroupPersistence groupPersistence) {
-        this.groupPersistence = groupPersistence;
-    }
-
-    /**
-     * Returns the resource local service.
-     *
-     * @return the resource local service
-     */
-    public com.liferay.portal.service.ResourceLocalService getResourceLocalService() {
-        return resourceLocalService;
-    }
-
-    /**
-     * Sets the resource local service.
-     *
-     * @param resourceLocalService the resource local service
-     */
-    public void setResourceLocalService(
-        com.liferay.portal.service.ResourceLocalService resourceLocalService) {
-        this.resourceLocalService = resourceLocalService;
-    }
-
-    /**
-     * Returns the user local service.
-     *
-     * @return the user local service
-     */
-    public com.liferay.portal.service.UserLocalService getUserLocalService() {
-        return userLocalService;
-    }
-
-    /**
-     * Sets the user local service.
-     *
-     * @param userLocalService the user local service
-     */
-    public void setUserLocalService(
-        com.liferay.portal.service.UserLocalService userLocalService) {
-        this.userLocalService = userLocalService;
-    }
-
-    /**
-     * Returns the user remote service.
-     *
-     * @return the user remote service
-     */
-    public com.liferay.portal.service.UserService getUserService() {
-        return userService;
-    }
-
-    /**
-     * Sets the user remote service.
-     *
-     * @param userService the user remote service
-     */
-    public void setUserService(
-        com.liferay.portal.service.UserService userService) {
-        this.userService = userService;
-    }
-
-    /**
-     * Returns the user persistence.
-     *
-     * @return the user persistence
-     */
-    public UserPersistence getUserPersistence() {
-        return userPersistence;
-    }
-
-    /**
-     * Sets the user persistence.
-     *
-     * @param userPersistence the user persistence
-     */
-    public void setUserPersistence(UserPersistence userPersistence) {
-        this.userPersistence = userPersistence;
-    }
-
-    /**
-     * Returns the d d l record set local service.
-     *
-     * @return the d d l record set local service
-     */
-    public com.liferay.portlet.dynamicdatalists.service.DDLRecordSetLocalService getDDLRecordSetLocalService() {
-        return ddlRecordSetLocalService;
-    }
-
-    /**
-     * Sets the d d l record set local service.
-     *
-     * @param ddlRecordSetLocalService the d d l record set local service
-     */
-    public void setDDLRecordSetLocalService(
-        com.liferay.portlet.dynamicdatalists.service.DDLRecordSetLocalService ddlRecordSetLocalService) {
-        this.ddlRecordSetLocalService = ddlRecordSetLocalService;
-    }
-
-    /**
-     * Returns the d d l record set remote service.
-     *
-     * @return the d d l record set remote service
-     */
-    public com.liferay.portlet.dynamicdatalists.service.DDLRecordSetService getDDLRecordSetService() {
-        return ddlRecordSetService;
-    }
-
-    /**
-     * Sets the d d l record set remote service.
-     *
-     * @param ddlRecordSetService the d d l record set remote service
-     */
-    public void setDDLRecordSetService(
-        com.liferay.portlet.dynamicdatalists.service.DDLRecordSetService ddlRecordSetService) {
-        this.ddlRecordSetService = ddlRecordSetService;
-    }
-
-    /**
-     * Returns the d d l record set persistence.
-     *
-     * @return the d d l record set persistence
-     */
-    public DDLRecordSetPersistence getDDLRecordSetPersistence() {
-        return ddlRecordSetPersistence;
-    }
-
-    /**
-     * Sets the d d l record set persistence.
-     *
-     * @param ddlRecordSetPersistence the d d l record set persistence
-     */
-    public void setDDLRecordSetPersistence(
-        DDLRecordSetPersistence ddlRecordSetPersistence) {
-        this.ddlRecordSetPersistence = ddlRecordSetPersistence;
-    }
-
-    /**
-     * Returns the d d m structure local service.
-     *
-     * @return the d d m structure local service
-     */
-    public com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalService getDDMStructureLocalService() {
-        return ddmStructureLocalService;
-    }
-
-    /**
-     * Sets the d d m structure local service.
-     *
-     * @param ddmStructureLocalService the d d m structure local service
-     */
-    public void setDDMStructureLocalService(
-        com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalService ddmStructureLocalService) {
-        this.ddmStructureLocalService = ddmStructureLocalService;
-    }
-
-    /**
-     * Returns the d d m structure remote service.
-     *
-     * @return the d d m structure remote service
-     */
-    public com.liferay.portlet.dynamicdatamapping.service.DDMStructureService getDDMStructureService() {
-        return ddmStructureService;
-    }
-
-    /**
-     * Sets the d d m structure remote service.
-     *
-     * @param ddmStructureService the d d m structure remote service
-     */
-    public void setDDMStructureService(
-        com.liferay.portlet.dynamicdatamapping.service.DDMStructureService ddmStructureService) {
-        this.ddmStructureService = ddmStructureService;
-    }
-
-    /**
-     * Returns the d d m structure persistence.
-     *
-     * @return the d d m structure persistence
-     */
-    public DDMStructurePersistence getDDMStructurePersistence() {
-        return ddmStructurePersistence;
-    }
-
-    /**
-     * Sets the d d m structure persistence.
-     *
-     * @param ddmStructurePersistence the d d m structure persistence
-     */
-    public void setDDMStructurePersistence(
-        DDMStructurePersistence ddmStructurePersistence) {
-        this.ddmStructurePersistence = ddmStructurePersistence;
-    }
-
-    /**
-     * Returns the journal article local service.
-     *
-     * @return the journal article local service
-     */
-    public com.liferay.portlet.journal.service.JournalArticleLocalService getJournalArticleLocalService() {
-        return journalArticleLocalService;
-    }
-
-    /**
-     * Sets the journal article local service.
-     *
-     * @param journalArticleLocalService the journal article local service
-     */
-    public void setJournalArticleLocalService(
-        com.liferay.portlet.journal.service.JournalArticleLocalService journalArticleLocalService) {
-        this.journalArticleLocalService = journalArticleLocalService;
-    }
-
-    /**
-     * Returns the journal article remote service.
-     *
-     * @return the journal article remote service
-     */
-    public com.liferay.portlet.journal.service.JournalArticleService getJournalArticleService() {
-        return journalArticleService;
-    }
-
-    /**
-     * Sets the journal article remote service.
-     *
-     * @param journalArticleService the journal article remote service
-     */
-    public void setJournalArticleService(
-        com.liferay.portlet.journal.service.JournalArticleService journalArticleService) {
-        this.journalArticleService = journalArticleService;
-    }
-
-    /**
-     * Returns the journal article persistence.
-     *
-     * @return the journal article persistence
-     */
-    public JournalArticlePersistence getJournalArticlePersistence() {
-        return journalArticlePersistence;
-    }
-
-    /**
-     * Sets the journal article persistence.
-     *
-     * @param journalArticlePersistence the journal article persistence
-     */
-    public void setJournalArticlePersistence(
-        JournalArticlePersistence journalArticlePersistence) {
-        this.journalArticlePersistence = journalArticlePersistence;
-    }
-
-    public void afterPropertiesSet() {
-        Class<?> clazz = getClass();
-
-        _classLoader = clazz.getClassLoader();
-    }
-
-    public void destroy() {
-    }
-
-    /**
-     * Returns the Spring bean ID for this bean.
-     *
-     * @return the Spring bean ID for this bean
-     */
-    @Override
-    public String getBeanIdentifier() {
-        return _beanIdentifier;
-    }
-
-    /**
-     * Sets the Spring bean ID for this bean.
-     *
-     * @param beanIdentifier the Spring bean ID for this bean
-     */
-    @Override
-    public void setBeanIdentifier(String beanIdentifier) {
-        _beanIdentifier = beanIdentifier;
-    }
-
-    @Override
-    public Object invokeMethod(String name, String[] parameterTypes,
-        Object[] arguments) throws Throwable {
-        Thread currentThread = Thread.currentThread();
-
-        ClassLoader contextClassLoader = currentThread.getContextClassLoader();
-
-        if (contextClassLoader != _classLoader) {
-            currentThread.setContextClassLoader(_classLoader);
-        }
-
-        try {
-            return _clpInvoker.invokeMethod(name, parameterTypes, arguments);
-        } finally {
-            if (contextClassLoader != _classLoader) {
-                currentThread.setContextClassLoader(contextClassLoader);
-            }
-        }
-    }
-
-    /**
-     * Performs an SQL query.
-     *
-     * @param sql the sql query
-     */
-    protected void runSQL(String sql) throws SystemException {
-        try {
-            DataSource dataSource = InfrastructureUtil.getDataSource();
-
-            SqlUpdate sqlUpdate = SqlUpdateFactoryUtil.getSqlUpdate(dataSource,
-                    sql, new int[0]);
-
-            sqlUpdate.update();
-        } catch (Exception e) {
-            throw new SystemException(e);
-        }
-    }
-}
