--- src/main/java/com/liferay/skinny/service/messaging/ClpMessageListener.java
@@ -1,25 +1,0 @@
-package com.liferay.skinny.service.messaging;
-
-import com.liferay.portal.kernel.messaging.BaseMessageListener;
-import com.liferay.portal.kernel.messaging.Message;
-
-import com.liferay.skinny.service.ClpSerializer;
-import com.liferay.skinny.service.SkinnyServiceUtil;
-
-
-public class ClpMessageListener extends BaseMessageListener {
-    public static String getServletContextName() {
-        return ClpSerializer.getServletContextName();
-    }
-
-    @Override
-    protected void doReceive(Message message) throws Exception {
-        String command = message.getString("command");
-        String servletContextName = message.getString("servletContextName");
-
-        if (command.equals("undeploy") &&
-                servletContextName.equals(getServletContextName())) {
-            SkinnyServiceUtil.clearService();
-        }
-    }
-}
