--- src/main/resources/META-INF/portlet-hbm.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN" "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
-
-<hibernate-mapping default-lazy="false" auto-import="false">
-	<import class="com.liferay.skinny.model.Skinny" />
-</hibernate-mapping>
