--- src/main/resources/META-INF/portlet-orm.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-
-<entity-mappings xmlns="http://java.sun.com/xml/ns/persistence/orm" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="1.0" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence/orm http://java.sun.com/xml/ns/persistence/orm_1_0.xsd">
-	<persistence-unit-metadata>
-		<xml-mapping-metadata-complete/>
-		<persistence-unit-defaults>
-			<access>PROPERTY</access>
-		</persistence-unit-defaults>
-	</persistence-unit-metadata>
-</entity-mappings>
