--- src/main/resources/META-INF/portlet-spring.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0"?>
-
-<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" default-destroy-method="destroy" default-init-method="afterPropertiesSet" xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
-	<bean id="com.liferay.skinny.service.SkinnyService" class="com.liferay.skinny.service.impl.SkinnyServiceImpl" />
-</beans>
