--- src/main/resources/service.properties
@@ -1,41 +1,0 @@
-##
-## Properties Override
-##
-
-    #
-    # Specify where to get the overridden properties. Updates should not be made
-    # on this file but on the overridden version of this file.
-    #
-    include-and-override=${base.path}/service-ext.properties
-
-##
-## Build
-##
-
-    build.namespace=Skinny
-    build.number=5
-    build.date=1455888060005
-    build.auto.upgrade=true
-
-##
-## Spring
-##
-
-    #
-    # Input a list of comma delimited Spring configurations. These will be
-    # loaded after the bean definitions specified in the
-    # portalContextConfigLocation parameter in web.xml.
-    #
-    spring.configs=\
-        WEB-INF/classes/META-INF/base-spring.xml,\
-        \
-        WEB-INF/classes/META-INF/hibernate-spring.xml,\
-        WEB-INF/classes/META-INF/infrastructure-spring.xml,\
-        \
-        WEB-INF/classes/META-INF/cluster-spring.xml,\
-        \
-        WEB-INF/classes/META-INF/portlet-spring.xml,\
-        \
-        WEB-INF/classes/META-INF/shard-data-source-spring.xml,\
-        \
-        WEB-INF/classes/META-INF/ext-spring.xml
