--- src/main/resources/service.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE service-builder PUBLIC "-//Liferay//DTD Service Builder 6.2.0//EN" "http://www.liferay.com/dtd/liferay-service-builder_6_2_0.dtd">
-
-<service-builder package-path="com.liferay.skinny">
-	<namespace>Skinny</namespace>
-	<entity name="Skinny" local-service="false" remote-service="true">
-		<reference package-path="com.liferay.dynamic.data.lists" entity="DDLRecordSet" />
-		<reference package-path="com.liferay.dynamic.data.mapping" entity="DDMStructure" />
-		<reference package-path="com.liferay.journal" entity="JournalArticle" />
-        <reference package-path="com.liferay.portal" entity="Group" />
-	</entity>
-</service-builder>
