--- src/main/webapp/WEB-INF/liferay-plugin-package.properties
@@ -1,11 +1,0 @@
-name=Skinny JSON Provider
-module-group-id=liferay
-module-incremental-version=1
-tags=content, ddl, json, rest, servicebuilder, services web plugin
-short-description=
-long-description=Liferay's Skinny JSON Provider app provides streamlined representations of your portal's Web Content Articles and Dynamic Data Lists. It gives you lean essential asset data that you can quickly process and analyze.<br />On installing the app, its JSON web services are available. You can invoke these services from any client that supports JSON. Liferay's page for the JSON web service API has forms that let you try the services. This page also shows you what the client code looks like in JavaScript, cURL, and URL formats.<br />That's the "skinny" on Liferay's Skinny JSON Provider!
-change-log=
-page-url=http://www.liferay.com
-author=Liferay, Inc.
-licenses=LGPL
-liferay-versions=7.0.0+,7.1.0+
