--- src/main/webapp/WEB-INF/web.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-
-<web-app version="2.4" xmlns="http://java.sun.com/xml/ns/j2ee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd">
-	<servlet>
-		<servlet-name>JSON Web Service Servlet</servlet-name>
-		<servlet-class>com.liferay.portal.kernel.servlet.PortalClassLoaderServlet</servlet-class>
-		<init-param>
-			<param-name>servlet-class</param-name>
-			<param-value>com.liferay.portal.jsonwebservice.JSONWebServiceServlet</param-value>
-		</init-param>
-		<load-on-startup>0</load-on-startup>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>JSON Web Service Servlet</servlet-name>
-		<url-pattern>/api/jsonws/*</url-pattern>
-	</servlet-mapping>
-</web-app>
